3.1 upload connector (#132)
* new page monitoring
* UpLoad Jar from Content page
* Move to 8.3 version
* Add a banner
* Produce the same element-template
* OAuth integration
* Organize import
* Improve interface to show the connection
* Remove Ping in the list on demand
* Update information any time
* Rename application and remove mustache

+++ src/main/java/io/camunda/cherry/runtime/HistoryFactory.java
@@ -49,7 +49,7 @@ public class HistoryFactory {
 
     HistoryPerformance.IntervalRule intervalRule = historyPerformance.getIntervalRuleByPeriod(periodStatistic);
     LocalDateTime dateThreshold = LocalDateTime.now();
-    dateThreshold = dateThreshold.minusMinutes(intervalRule.intervalInMinutes * intervalRule.numberOfIntervals);
+    dateThreshold = dateThreshold.minusMinutes((long) intervalRule.intervalInMinutes * intervalRule.numberOfIntervals);
 
     List<Map<String, Object>> listStats = runnerExecutionRepository.selectStatusStats(runnerType, dateThreshold);
     for (Map<String, Object> recordStats : listStats) {
